move the test to the latest test suit
